user-airdrop/requirements.txt
pycoingecko
etherscan-python
